Change utf8 characters to ascii.
The dep5 file contains utf-8 characters.
This is in general not an issue but can lead to issues if other
programs do not correctly handle utf8.
Unfortunately, the reuse tool does not handle utf8 characters
correctly. This can create a showstopper on windows.

The utf-8 characters in the dep5 file seem to be caused by
some word export where some special quotation marks were used.
We can simply replace them by regular ascii characters. This makes
the file both compatible to ascii and utf-8.

I will also send a fix to the reuse tool, but making the dep5 file
ascii compatible is still desirable as it allows to proceed now.